feat: add css var display value
