+++ utils/generate/src/types.ts
@@ -24,4 +24,5 @@ export interface FontObject {
 	unicodeRange?: string;
 	comment?: string;
 	spacer?: string;
+	displayVar?: boolean;
 }
